vfs_glusterfs: format acls and xattr correctly
1. In case of setting acls
The format to set acls using xattr by setxattr call
prescribes the acl entries to be in little endian format.
This patch converts the values to little endian.

2. In case of getting acls
As acl entries are always stored in little endian byte order,
we have to convert the data to cpu's native byte order before
using it.

Signed-off-by: Raghavendra Talur <rtalur@redhat.com>